Disable CGO for CI Builds
Currently releases contain a linux/amd64 skupper-cli asset built with
CGO, the default behavior for the go toolchain for native builds. The
consequence of this is a less portable executable. This change disables
CGO for our tests and builds in CircleCI, and will produce statically
linked pure go executables for all platforms uniformly.

Signed-off-by: c-kruse <ctkruse99@gmail.com>