Refactor configuration source for sslProfiles
Adds the internal/sslprofile and internal/kube/secrets packages to serve
as the source of sslProfile configuration. The internal/sslprofile
package works with the internal/site package to produce router
configuration for RouterAccess and Link resources. The
internal/kube/secrets package works with the kube controller and
kube-adaptor executables to coordinate sslProfiles with kubernetes
Secret resources.

Signed-off-by: Christian Kruse <christian@c-kruse.com>