Implements Certificate Rotation Fundamentals
Adds a new internal/kube/secrets package that implements the logic for
watching kubernetes Secret resources containing the credentials to be
used in router sslProfile configuration, and automatically invoking the
certificate rotation primitives in router configuration when Secret data
changes.

Introduces a new partially backwards incompatible interface between the
controller and kube-adaptor in order to accurately associate a given
Secret Resource revision with an sslProfile+ordinal pair. Prior versions
of the kube-adaptor should work with the updated controller logic, but
updated kube-adaptors will wait for an updated controller before
starting.

Signed-off-by: Christian Kruse <christian@c-kruse.com>